[Create] UserDetailsDTO, UserDetailsServiceImpl [Improvement] DTOs now have non null id (only entities need a nullable one)

+++ src/main/kotlin/it/polito/wa2/lab2/domain/Transaction.kt
@@ -34,8 +34,8 @@ final class Transaction(
 fun Transaction.toDTO(): TransactionDTO =
     TransactionDTO(
         amount,
-        from.getId(),
-        to.getId(),
+        from.getId() as Long,
+        to.getId() as Long,
         timestamp,
-        getId()
+        getId() as Long
     )
